move factory helper into factories file
